Create TabBarRemoteMessagePresentable
